feat: use new nom based parser
